requirements.txt
python-socketio[client]==4.6.1
load_dotenv
dataclasses-json==0.5.7